LPS-148350 Rename package
